Added a loop
In this edit, I added a loop to print each pokemon in my array. If the pokemon was less that 6 feet tall, it should print "thats pretty small!" and if it was greater than or equal to 6 feet tall, it should print "thats quite large". I also removed the word "foot" from the array itself since the loop would only recognize the number and not the "foot" or "feet" part of the height. I instead added this to the loop to print "feet".