Changed one array object and conditional
I changed one of the pokemon (Garchomp to Sceptile) in the array to have just one type for now. I kept the height as 6 so that the third array object would still print "feet tall, thats quite large!" I also edited the conditional for the "for" loop so that the code would stop after the third pokemon in the array. Lastly, I changed the height maximum for the "else if" portion of the "for" loop just to make it easier to understand.