Created addListItem and event listener in IIFE
I created an addListItem function in order to add a button to each pokemon in my list. I also added an event listener so that each time a pokemon is clicked, it logs the information to the console. I styled the page with CSS by adding the pokemon logo, a title, styled the buttons and added Electrode to the background of the page.