Pokedex API added
I deleted my array of 6 pokemon. I also added the pokemon API to load the list of pokemon and details about the pokemon so that they will be logged to the console when clicked. I also adjusted my CSS to make the app more appealing to the user by making the background a gradient, changing the color of the buttons to a more transluscent grey color, and chaned the color of the text inside each button.