Modal Display picture and type of pokemon
Made adjustments to code by making sure load details was called before the show modal function so that the details of the pokemon are loaded before displaying the modal so that the modal actually displays correctly. I also made three new variables according to the new classes made from the bootstrap HTML and attached the name, height, types and images to the title and body of the modal.